Introduced Participants API
Introduced Live Participants API to Communicate with Django Server
