+++ src/constants/participantsConsts.js
@@ -1,3 +1,4 @@
 export const REVIEWED_STATUS_NOT_REVIEWED = 'NOT REVIEWED';
 export const DEFAULT_PROFILE_IMAGE_URL = 'https://images.pexels.com/photos/1194806/pexels-photo-1194806.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260';
 export const ERIK_LINKEDIN_PROFILE_URL = 'https://www.linkedin.com/in/erik-butcher?lipi=urn%3Ali%3Apage%3Ad_flagship3_profile_view_base_contact_details%3BaQmDX3xERc6qMWY5Y%2BmFBw%3D%3D';
+export const DEFAULT_DJANGO_BASE_URL = 'http://localhost:8000';
